Feature/swap

+++ src/config/seedlogs.js
@@ -1,52 +1,35 @@
-import mongoose from 'mongoose';
-
-// Create logSchema
-const logSchema = new mongoose.Schema({
-  logId: {
-    type: String,
-    required: true,
-    unique: true,
-  },
-  action: {
-    type: String,
-    required: true,
-  },
-  timestamp: {
-    type: Date,
-    required: true,
-    default: Date.now,
-  },
-  userId: {
-    type: String,
-    required: false,
-  },
-});
-
-const Log = mongoose.model('Log', logSchema);
+const Log = require('../models/Log');
+const mongoose = require('mongoose');
 
 const seedLogs = async () => {
-  const logs = [
-    {
-      logId: new mongoose.Types.ObjectId().toString(),
-      action: 'user_created',
-      timestamp: new Date(),
-      userId: '1',
-    },
-    {
-      logId: new mongoose.Types.ObjectId().toString(),
-      action: 'item_created',
-      timestamp: new Date(),
-      userId: '2',
-    },
-    {
-      logId: new mongoose.Types.ObjectId().toString(),
-      action: 'swap_created',
-      timestamp: new Date(),
-      userId: '3',
-    },
-  ];
-  await Log.deleteMany({});
-  await Log.insertMany(logs);
+  // Only seed if there are no logs
+  const logCount = await Log.countDocuments();
+  if (logCount === 0) {
+    console.log('Seeding logs...');
+
+    const logs = [
+      {
+        logId: new mongoose.Types.ObjectId().toString(),
+        action: 'user_created',
+        timestamp: new Date(),
+        userId: '1',
+      },
+      {
+        logId: new mongoose.Types.ObjectId().toString(),
+        action: 'item_created',
+        timestamp: new Date(),
+        userId: '2',
+      },
+      {
+        logId: new mongoose.Types.ObjectId().toString(),
+        action: 'swap_created',
+        timestamp: new Date(),
+        userId: '3',
+      },
+    ];
+    await Log.deleteMany({});
+    await Log.insertMany(logs);
+  };
 };
 
-export default seedLogs;
+module.exports = seedLogs;
